correctos dtos y resource Feria, test de resurce feria y pabellon

+++ artesanias-01-web/src/main/java/co/edu/uniandes/csw/artesanias/resources/FeriaArtesanalResource.java
@@ -76,7 +76,7 @@ public class FeriaArtesanalResource {
     
     @DELETE
     @Path("{id: \\d+}")
-    public void deleteFeria(@PathParam("id") Integer id){
+    public void deleteFeria(@PathParam("id") Long id){
         logica.deleteFeria(id);
     }
 }
